+++ template_variables/src/epilot_template_variables/models/customvariablessearchparams.py
@@ -13,6 +13,7 @@ class CustomVariablesSearchParamsType(str, Enum):
 
     ORDER_TABLE = "order_table"
     CUSTOM = "custom"
+    JOURNEY_LINK = "journey_link"
 
 
 class CustomVariablesSearchParamsTypedDict(TypedDict):